Modify style
Show auth card if user isn't signed in, add some contrast
