+++ src/components/FirebaseImage.tsx
@@ -1,5 +1,5 @@
 import { getDownloadURL, ref } from "firebase/storage";
-import { useEffect, useState } from "react";
+import { useMemo, useState } from "react";
 import { ImageProps } from "react-bootstrap";
 import { storage } from "../config/firebase";
 import Image from "react-bootstrap/Image";
@@ -7,7 +7,7 @@ import Image from "react-bootstrap/Image";
 export const FirebaseImage = ({ ...props }: ImageProps) => {
   const [imageSrc, setImageSrc] = useState("");
 
-  useEffect(() => {
+  useMemo(() => {
     getDownloadURL(ref(storage, props.src)).then((val) => {
       setImageSrc(val);
     });
